requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
groq==0.11.0
httpx==0.27.0
python-dotenv==1.0.0
pymongo==4.6.1
